% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{calc_risk_term}
\alias{calc_risk_term}
\title{Calculate absolute risk profiles after non-terminal event}
\usage{
calc_risk_term(
  para,
  Xmat3,
  hazard,
  knots_list = NULL,
  t_cutoff,
  t_start,
  tol = 0.001,
  frailty = TRUE,
  type = "marginal",
  gamma = 1,
  model = "semi-markov",
  h3_tv = "none",
  tv_knots = NULL
)
}
\arguments{
\item{para}{A numeric vector of parameters, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the \eqn{k_1+k_2+k_3+1} element corresponds to the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.}

\item{Xmat3}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}

\item{t_start}{Numeric scalar indicating the dynamic start time to compute the risk profile. Set to 0 by default.}

\item{tol}{Numeric value for the tolerance of the numerical integration procedure.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to TRUE.}

\item{type}{String either indicating 'marginal' for population-averaged probabilities,
or 'conditional' for probabilities computed at the specified gamma}

\item{gamma}{Numeric value indicating the fixed level of the frailty assumed for predicted probabilities,
if 'type' is set to 'conditional'}

\item{model}{String specifying the transition assumption}

\item{h3_tv}{String indicating whether there is an effect of t1 on hazard 3.}

\item{tv_knots}{for piecewise effect of t1 in h3, these are the knots at which the effect jumps}
}
\value{
if Xmat has only one row, and t_cutoff is a scalar, then returns a 4 element row matrix
of probabilities. If Xmat has \code{n} rows, then returns an \code{n} by 4 matrix of probabilities.
If Xmat has \code{n} rows and t_cutoff is a vector of length \code{s}, then returns an \code{s} by 4 by \code{n} array.
}
\description{
This function calculates absolute risk profiles conditional on non-terminal
event already occurring.
}
