% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute AUC for Non-terminal and Terminal outcomes}
\usage{
compute_auc(dat, t_cutoff, pred_mat)
}
\arguments{
\item{dat}{Data}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}

\item{pred_mat}{Output from calc_risks function}
}
\value{
a vector with the non-terminal and terminal AUC.
}
\description{
Function to compute univariate prediction statistics
}
