% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{compute_ccp}
\alias{compute_ccp}
\title{Compute IPCW-adjusted Correct Classification Probability}
\usage{
compute_ccp(outcome_mat, pred_mat, ipcw_mat)
}
\arguments{
\item{outcome_mat}{Output from get_outcome_mat function}

\item{pred_mat}{Output from calc_risks function}

\item{ipcw_mat}{Output from get_ipcw_mat function}
}
\value{
a scalar CCP
}
\description{
This function computes the IPCW-adjusted Correct Classification Probability. Presented in Li (2019)
}
