% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{compute_hum}
\alias{compute_hum}
\title{Compute IPCW-adjusted Hypervolume Under the Manifold}
\usage{
compute_hum(outcome_mat, pred_mat, ipcw_mat)
}
\arguments{
\item{outcome_mat}{Output from get_outcome_mat function}

\item{pred_mat}{Output from calc_risks function}

\item{ipcw_mat}{Output from get_ipcw_mat function}
}
\value{
a scalar
}
\description{
This function computes the IPCW-adjusted Hypervolume under the manifold, as in Lee (unpublished).
}
