% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{compute_nri}
\alias{compute_nri}
\title{Compute IPCW-adjusted Net Reclassification Improvement}
\usage{
compute_nri(outcome_mat, pred_mat1, pred_mat2, ipcw_mat)
}
\arguments{
\item{outcome_mat}{Output from get_outcome_mat function}

\item{pred_mat1}{matrix of predicted probabilities with model 1 ("old" model)}

\item{pred_mat2}{matrix of predicted probabilities with model 2 ("new" model)}

\item{ipcw_mat}{Output from get_ipcw_mat function}
}
\value{
a scalar NRI
}
\description{
This function computes the IPCW-adjusted Net Reclassification Probability. Presented as S in Li (2013) and Wang (2020)
}
