% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{compute_pdi}
\alias{compute_pdi}
\title{Compute IPCW-adjusted Polytomous Discrimination Index}
\usage{
compute_pdi(outcome_mat, pred_mat, ipcw_mat)
}
\arguments{
\item{outcome_mat}{Output from get_outcome_mat function}

\item{pred_mat}{Output from calc_risks function}

\item{ipcw_mat}{Output from get_ipcw_mat function}
}
\value{
a scalar HUM
}
\description{
This function computes the IPCW-adjusted polytomous discrimination index. Proposed in Van Caster and presented in Li
}
