% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{compute_score}
\alias{compute_score}
\title{Compute prediction performance score}
\usage{
compute_score(outcome_mat, pred_mat, ipcw_mat, score = "brier")
}
\arguments{
\item{outcome_mat}{Output from get_outcome_mat function}

\item{pred_mat}{Output from calc_risks function}

\item{ipcw_mat}{Output from get_ipcw_mat function}

\item{score}{String indicating whether 'brier' score, or 'entropy' should be computed.}
}
\value{
a vector.
}
\description{
This function takes in all of the ingredients needed for prediction validation,
and returns the corresponding scores.
}
