% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_mat_functions.R
\name{contrast_mat_list}
\alias{contrast_mat_list}
\title{Create List of Contrast Matrices for Parameter Fusion}
\usage{
contrast_mat_list(
  nP0,
  nP1,
  nP2,
  nP3,
  penalty_fusedcoef,
  lambda_fusedcoef,
  penalty_fusedbaseline,
  lambda_fusedbaseline,
  hazard,
  penweights_list
)
}
\value{
Returns a list with elements named as above, each containing a numeric contrast matrix.
}
\description{
Creates list of numeric matrices corresponding to the fusion of blocks of parameters.
The list is indexed by pairs of parameter blocks being fused, i.e.,
\itemize{
\item \code{fusedcoef12},\code{fusedcoef13},\code{fusedcoef23} list elements correspond to
matrices that, when multiplied to the parameter vector, yield a vector of fused differences
between corresponding regression parameters between the stated transition hazards.
\item \code{fusedbaseline12},\code{fusedbaseline13},\code{fusedbaseline23} list elements
are as above, but fusing the baseline hazard parameters.
}
}
