% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_validate.R
\name{cv_predict}
\alias{cv_predict}
\title{Calculate Cross-validated prediction performance metrics}
\usage{
cv_predict(
  Formula,
  data,
  na.action = "na.fail",
  subset = NULL,
  hazard = c("weibull"),
  frailty = TRUE,
  model,
  knots_list = NULL,
  p0_vec = rep(4, 3),
  startVals = NULL,
  optim_method = if (tolower(hazard) \%in\% c("royston-parmar", "rp")) "BFGS" else
    "L-BFGS-B",
  n_folds,
  t_cutoff,
  t_start = 0,
  tol = 0.001,
  type = "marginal",
  gamma = 1,
  h3_tv,
  tv_knots,
  verbose = 0
)
}
\arguments{
\item{Formula}{a Formula object, with the outcome on the left of a
\code{~}, and covariates on the right. It is of the form, \code{time to non-terminal
  event + corresponding censoring indicator | time to terminal event
  + corresponding censoring indicator ~ covariates for h_1 |
  covariates for h_2 | covariates for h_3}. For example, \code{y_1 + delta_1 | y_2 + delta_2 ~ x_1 | x_2 | x_3}.}

\item{data}{a \code{data.frame} in which to interpret the variables named in Formula.}

\item{na.action}{how NAs are treated. See \code{\link[stats]{model.frame}}.}

\item{subset}{a specification of the rows to be used: defaults to all rows. See \code{\link[stats]{model.frame}}.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to TRUE.}

\item{model}{String specifying the transition assumption}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard. If
\code{NULL}, will be created by \code{\link{get_default_knots_list}}.}

\item{p0_vec}{vector of length three of integers indicating how many baseline hazard parameters
should be specified for each of the three transition hazards. This input is only relevant when
hazard is something other than "weibull" and is superceded by knots_list.}

\item{startVals}{A numeric vector of parameter starting values, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the \eqn{k_1+k_2+k_3+1} element corresponds to the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.
If set to \code{NULL}, will be generated automatically using \code{\link{get_start}}.}

\item{optim_method}{a string naming which \code{optim} method should be used.}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}

\item{t_start}{Numeric scalar indicating the dynamic start time to compute the risk profile. Set to 0 by default.}

\item{tol}{Numeric value for the tolerance of the numerical integration procedure.}

\item{type}{String either indicating 'marginal' for population-averaged probabilities,
or 'conditional' for probabilities computed at the specified gamma}

\item{gamma}{Numeric value indicating the fixed level of the frailty assumed for predicted probabilities,
if 'type' is set to 'conditional'}

\item{h3_tv}{String indicating whether there is an effect of t1 on hazard 3.}

\item{tv_knots}{for piecewise effect of t1 in h3, these are the knots at which the effect jumps}
}
\value{
if Xmat has only one row, and t_cutoff is a scalar, then returns a 4 element row matrix
of probabilities. If Xmat has \code{n} rows, then returns an \code{n} by 4 matrix of probabilities.
If Xmat has \code{n} rows and t_cutoff is a vector of length \code{s}, then returns an \code{s} by 4 by \code{n} array.
}
\description{
This function calculates cross-validated prediction performance metrics
}
