% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_validate.R
\name{cv_validate}
\alias{cv_validate}
\title{Calculate Cross-validated prediction performance metrics}
\usage{
cv_validate(
  cv_predict_out_list,
  metrics = c("brier", "auc", "hum", "ccp", "pdi"),
  verbose = 0
)
}
\arguments{
\item{cv_predict_out_list}{is an object output from \code{cv_predict} function}

\item{metrics}{is a vector of strings naming what metrics to compute. Options are "brier", "auc", "hum", "ccp",and "pdi".}

\item{verbose}{Numeric indicating the amount of intermediate information that should be printed during the cross-validation process.
Larger numbers correspond to more printed information.}
}
\value{
if Xmat has only one row, and t_cutoff is a scalar, then returns a 4 element row matrix
of probabilities. If Xmat has \code{n} rows, then returns an \code{n} by 4 matrix of probabilities.
If Xmat has \code{n} rows and t_cutoff is a vector of length \code{s}, then returns an \code{s} by 4 by \code{n} array.
}
\description{
This function calculates cross-validated prediction performance metrics
}
