% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection.R
\name{forward_selection}
\alias{forward_selection}
\title{Forward Selection Procedure for Parametric Illness-Death Model}
\usage{
forward_selection(
  vars,
  data,
  na.action = "na.fail",
  subset = NULL,
  hazard = c("weibull"),
  frailty = TRUE,
  model,
  knots_list = NULL,
  select_crit = "bic",
  verbose = 0,
  control = NULL
)
}
\arguments{
\item{vars}{character string with all of the possible variable names to be searched through.}

\item{data}{a \code{data.frame} in which to interpret the variables named in Formula.}

\item{na.action}{how NAs are treated. See \code{\link[stats]{model.frame}}.}

\item{subset}{a specification of the rows to be used: defaults to all rows. See \code{\link[stats]{model.frame}}.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to TRUE.}

\item{model}{String specifying the transition assumption}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard. If
\code{NULL}, will be created by \code{\link{get_default_knots_list}}.}

\item{select_crit}{a string indicating what criterion should be used to determine whether a covariate should be added.}

\item{verbose}{Numeric indicating the amount of iteration information should be printed to the user.
Higher numbers provide more detailed information to user, but will slow down the algorithm.}

\item{control}{a list of control attributes passed directly into the \code{optim} function.}
}
\value{
A list.
}
\description{
This method is meant as a comparator for the penalized estimation procedure.
}
