% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_basis}
\alias{get_basis}
\title{Get Basis Function Values for Flexible Hazard Specifications}
\usage{
get_basis(x, knots, hazard, deriv = FALSE, flexsurv_compatible = FALSE)
}
\arguments{
\item{x}{Numeric vector of event times (e.g., \code{y1} or \code{y2}) at which
to generate basis function values.}

\item{knots}{Increasing vector of integers corresponding to the knots used
in the desired spline or piecewise specification. Often an element of
list generated from \code{\link{get_default_knots_list}}.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{deriv}{Boolean for whether returned values should be from derivatives of
basis functions if \code{TRUE}, or original basis functions if \code{FALSE}.}

\item{flexsurv_compatible}{For Royston-Parmar basis, boolean for whether returned values should be untransformed,
as in \code{flexsurv} package, if \code{TRUE}, or whether they should be transformed to remove correlation,
as in ns package otherwise.}
}
\value{
A matrix with each row corresponding to an element of the input, and
each column giving the corresponding basis function value.
}
\description{
Get Basis Function Values for Flexible Hazard Specifications
}
