% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_default_knots_list}
\alias{get_default_knots_list}
\title{Generate List of Knot Location Vectors from Event Times}
\usage{
get_default_knots_list(y1, y2, delta1, delta2, p01, p02, p03, hazard, model)
}
\arguments{
\item{y1}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{y2}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta1}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{p01, p02, p03}{Integers indicating how many parameters the model for each
transition baseline hazard should specify.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{model}{String specifying the transition assumption}
}
\value{
A list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
}
\description{
This function creates a list containing three numeric vectors. Each numeric vector
is a sequence of increasing integers giving the location of knots used for
spline and piecewise baseline hazard specifications. This function
generates this list according to the conventions and recommended locations
of these knots, which depends on the choice of hazard specification, number
of knots requested, and distribution of observed event times.
}
