% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risks.R
\name{get_outcome_mat}
\alias{get_outcome_mat}
\title{Get matrix of observed outcome categories}
\usage{
get_outcome_mat(y1, y2, delta1, delta2, t_cutoff)
}
\arguments{
\item{y1}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{y2}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta1}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}
}
\value{
a matrix or array.
}
\description{
This function returns a matrix giving the observed outcome categories of each observation at various
time cutoffs.
}
