% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_functions.R
\name{nlogLikPW_ID_frail_M}
\alias{nlogLikPW_ID_frail_M}
\title{Negative Log-Likelihood Function for Illness-Death Model}
\usage{
nlogLikPW_ID_frail_M(
  para,
  y1,
  y2,
  delta1,
  delta2,
  Xmat1 = NULL,
  Xmat2 = NULL,
  Xmat3 = NULL,
  basis1,
  basis2,
  basis3,
  basis3_y1
)
}
\arguments{
\item{para}{A numeric vector of parameters, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the \eqn{k_1+k_2+k_3+1} element corresponds to the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.}

\item{y1, y2}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta1, delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{Xmat1, Xmat2, Xmat3}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{basis1, basis2, basis3, basis3_y1}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function values at the corresponding \code{y1} and \code{y2} values.
Under semi-Markov model, basis3 represents basis derived from \eqn{y_2-y_1} and \code{basis3_y1} is unused,
while under Markov model, basis3 represents basis derived from \eqn{y_2} and \code{basis3_y1} is from \eqn{y_1}
Not used under Weibull model.}
}
\value{
Returns numeric sum of negative log likelihood contributions.
}
\description{
Function returning the negative log-likelihood for the illness-death model,
under piecewise constant baseline hazard, gamma subject-specific frailty,
and Markov transition assumption.
Typically, this function will not be used directly by the user, but as part of a
larger estimation procedure.
}
