% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simID_PW.R
\name{simID_PW}
\alias{simID_PW}
\title{The function that simulates independent/cluster-correlated semi-competing
risks data under semi-Markov PEM/PEM-MVN models.}
\usage{
simID_PW(
  id = NULL,
  x1,
  x2,
  x3,
  beta1.true,
  beta2.true,
  beta3.true,
  phi1.true,
  phi2.true,
  phi3.true,
  theta.true,
  SigmaV.true = NULL,
  cens,
  knots_list
)
}
\arguments{
\item{id}{A vector of cluster information for \code{n} subjects.
The cluster membership must be set to consecutive positive integers, \eqn{1:J}.
Required only when generating clustered data.}

\item{x1, x2, x3}{Covariate matrices with \code{n} rows.}

\item{beta1.true, beta2.true, beta3.true}{Vectors of true regression parameter values.
The length of each vector should equal the number of columns in the corresponding covariate matrix.}

\item{phi1.true, phi2.true, phi3.true}{Vectors of true baseline parameter values.}

\item{theta.true}{True value for \eqn{\theta}.}

\item{SigmaV.true}{True value for covariance matrix of MVN cluster-level random effects.
Required only when generating clustered data. Should be a numeric \eqn{J\times J} matrix.}

\item{cens}{A numeric vector of two elements. The right censoring times are generated from Uniform(\eqn{cens[1]}, \eqn{cens[2]}).}

\item{knots_list}{A list containing three numeric vectors, representing the breakpoints
of the piecewise specification for each transition hazard (excluding 0).}
}
\value{
returns a data.frame containing semi-competing risks outcomes from \code{n} subjects.
It is of dimension \eqn{n\times 4}: the columns correspond to \eqn{y_1}, \eqn{\delta_1}, \eqn{y_2}, \eqn{\delta_2}. \cr
\itemize{
\item{y1}{a vector of \code{n} times to the non-terminal event}
\item{y2}{a vector of \code{n} times to the terminal event}
\item{delta1}{a vector of \code{n} censoring indicators for the non-terminal event time (1=event occurred, 0=censored)}
\item{delta2}{a vector of \code{n} censoring indicators for the terminal event time (1=event occurred, 0=censored)}
}
}
\description{
The function that simulates independent/cluster-correlated semi-competing
risks data under semi-Markov PEM/PEM-MVN models.
}
